real-estate-analyzer/requirements.txt
# pip install -r requirements.txt 로 설치

#Flask & Web Server
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==0.21.1
firebase-admin==6.1.0
requests
gunicorn

#OCR & Image Processing
easyocr==1.7.1
opencv-python==4.9.0.80
scikit-image==0.22.0
numpy==1.26.4
Pillow==10.3.0

#AI Model
google-generativeai==0.3.2

# Data Analysis (시세계산용)
pandas
scikit-learn

# ================================================================
# 🧠 PyTorch (EasyOCR의 핵심 엔진) - CPU 버전
# GPU 관련(--extra-index-url) 옵션을 주석 처리하여,
# 배포 환경에 맞는 CPU 전용 PyTorch가 자동으로 설치되도록 합니다.
# ================================================================
# --extra-index-url https://download.pytorch.org/whl/cu126 이거는 서버용입니다.
# 배포할때 필요하니깐 주석처리 유지해주세요
torch
torchvision
torchaudio
